requirements.txt
# --- Core CLI and UX ---
typer>=0.9.0
rich>=13.7.0
colorama>=0.4.6
emoji>=2.12.1

# --- AI / LLM Integration ---
openai>=1.35.0
langchain>=0.2.6
langchain-community>=0.3.10
langchain-experimental>=0.0.62
langchainhub>=0.1.18
langsmith>=0.1.92
langchain-google-genai>=2.1.12
langchain-huggingface>=0.3.1
transformers>=4.57.0
accelerate>=1.10.0
bitsandbytes>=0.48.1
smolagent>=1.22.0
# --- Agent and Tools ---
duckduckgo-search>=5.3.1
ddgs>=9.6.0
wikipedia>=1.4.0
arxiv>=2.1.0
beautifulsoup4>=4.12.3
requests>=2.31.0

# --- Shell & System ---
pexpect>=4.9.0
psutil>=5.9.8
platformdirs>=4.3.0
fastapi>=0.111.0
uvicorn>=0.30.1
python-dotenv>=1.0.1

# --- Data & Models ---
pydantic>=2.7.1
numpy>=1.26.4
pandas>=2.2.2
tqdm>=4.66.4
pillow>=11.3.0

# --- Localization / i18n ---
babel>=2.14.0
polib>=1.2.0

# --- Logging & Utilities ---
loguru>=0.7.2
pyyaml>=6.0.1
filetype>=1.2.0

# --- Developer & Testing ---
black>=24.4.2
pytest>=8.2.2
ipython>=8.25.0
inquirer>=3.4.1
